Migrate Order

+++ app/Http/Controllers/ProductController.php
@@ -12,7 +12,14 @@ class ProductController extends Controller
 {
     public function index()
     {
-        $data = Product::where('company_id', Auth::user()->company_id)->with('category_product')->paginate(5);
+        // $data = Product::with(['category_product' => function ($query) {
+        //     $query->where('company_id', Auth::user()->company_id);
+        // }])->paginate(5);
+        $data = Product::select('products.*')
+        ->leftJoin('category_products', 'category_products.id', '=' , 'products.category_id')
+        ->where('company_id', Auth::user()->company_id)
+        ->paginate(5);
+
         return view('dashboard.master-data.product.index', [
             'data' => $data
         ]);
@@ -20,7 +27,7 @@ class ProductController extends Controller
 
     public function create()
     {
-        $lists = CategoryProduct::all();
+        $lists = CategoryProduct::where('company_id', Auth::user()->company_id)->get();
         return view('dashboard.master-data.product.create', ["lists" => $lists]);
     }
 
@@ -58,9 +65,19 @@ class ProductController extends Controller
 
     public function edit($id)
     {
-        $data = Product::findOrFail($id);
-        $lists = CategoryProduct::all();
-        return view('dashboard.master-data.product.edit', ["data" => $data, "lists" => $lists]);
+        // $data = Product::findOrFail($id);
+        $data = Product::select('products.*', 'company_id')
+        ->leftJoin('category_products', 'category_products.id', '=' , 'products.category_id')
+        ->where('company_id', Auth::user()->company_id)
+        ->where('products.id', $id)
+        ->first();
+
+        if ($data && $data->company_id == Auth::user()->company_id) { 
+            $lists = CategoryProduct::where('company_id', Auth::user()->company_id)->get();
+            return view('dashboard.master-data.product.edit', ["data" => $data, "lists" => $lists]);
+        } else {
+            return redirect()->route('dashboard.master-data.product')->with('failed', 'Oops! Looks like you followed a bad link. If you think this is a problem with us, please tell us.');
+        }
     }
 
     public function update(Request $request, $id)
@@ -72,35 +89,54 @@ class ProductController extends Controller
             'description' => 'required',
         ]);
 
-        $data = Product::findOrFail($id);
+        // $data = Product::findOrFail($id);
+        $data = Product::select('products.*', 'company_id')
+        ->leftJoin('category_products', 'category_products.id', '=' , 'products.category_id')
+        ->where('company_id', Auth::user()->company_id)
+        ->where('products.id', $id)
+        ->first();
 
-        if ($request->file('image')) {
-            if ($request->old_image) {
-                Storage::delete($request->old_image);
+        if ($data && $data->company_id == Auth::user()->company_id) { 
+            if ($request->file('image')) {
+                if ($request->old_image) {
+                    Storage::delete($request->old_image);
+                }
+                $validate['image'] = $request->file('image')->storeAs('images/master-data/product', time() . '.' . $request->image->extension());
             }
-            $validate['image'] = $request->file('image')->storeAs('images/master-data/product', time() . '.' . $request->image->extension());
-        }
-
-        $validate['is_available'] = $request["is_available"] ? 1 : 0;
-
-        $update = $data->update($validate);
-
-        if ($update) {
-            return redirect()->route('dashboard.master-data.product')->with('success', "Successfully to update  product");
+    
+            $validate['is_available'] = $request["is_available"] ? 1 : 0;
+    
+            $update = $data->update($validate);
+    
+            if ($update) {
+                return redirect()->route('dashboard.master-data.product')->with('success', "Successfully to update  product");
+            } else {
+                return redirect()->route('dashboard.master-data.product')->with('failed', "Failed to update  product");
+            }
         } else {
-            return redirect()->route('dashboard.master-data.product')->with('failed', "Failed to update  product");
+            return redirect()->route('dashboard.master-data.product')->with('failed', 'Oops! Looks like you followed a bad link. If you think this is a problem with us, please tell us.');
         }
     }
 
     public function destroy($id)
     {
-        $product = Product::find($id);
-        Storage::delete($product->image);
-        $delete =  Product::destroy($id);
-        if ($delete) {
-            return redirect()->route('dashboard.master-data.product')->with('success', "Successfully to delete  product");
+        // $product = Product::find($id);
+        $product = Product::select('products.*', 'company_id')
+        ->leftJoin('category_products', 'category_products.id', '=' , 'products.category_id')
+        ->where('company_id', Auth::user()->company_id)
+        ->where('products.id', $id)
+        ->first();
+
+        if ($product && $product->company_id == Auth::user()->company_id) { 
+            Storage::delete($product->image);
+            $delete =  Product::destroy($id);
+            if ($delete) {
+                return redirect()->route('dashboard.master-data.product')->with('success', "Successfully to delete  product");
+            } else {
+                return redirect()->route('dashboard.master-data.product')->with('failed', "Failed to delete  product");
+            }
         } else {
-            return redirect()->route('dashboard.master-data.product')->with('failed', "Failed to delete  product");
+            return redirect()->route('dashboard.master-data.product')->with('failed', 'Oops! Looks like you followed a bad link. If you think this is a problem with us, please tell us.');
         }
     }
 }
